requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
supabase
python-dotenv
pydantic
email-validator
